requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.2.1
PyMySQL==1.1.0
requests==2.32.4
resend==2.8.0
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.5.0
whitenoise==6.9.0
